Remove factory
