--- paymentsheet/src/test/java/com/stripe/android/paymentsheet/ui/FakeEditCardDetailsInteractorFactory.kt
@@ -1,18 +1,0 @@
-package com.stripe.android.paymentsheet.ui
-
-import com.stripe.android.model.PaymentMethod
-
-internal class FakeEditCardDetailsInteractorFactory(
-    private val shouldShowCardBrandDropdown: Boolean = true
-) : EditCardDetailsInteractor.Factory {
-    override fun create(
-        card: PaymentMethod.Card,
-        onCardUpdateParamsChanged: CardUpdateParamsCallback
-    ): EditCardDetailsInteractor {
-        return FakeEditCardDetailsInteractor(
-            card = card,
-            shouldShowCardBrandDropdown = shouldShowCardBrandDropdown,
-            onCardUpdateParamsChanged = onCardUpdateParamsChanged
-        )
-    }
-}
